+++ nmanga/cli/releases.py
@@ -114,6 +114,7 @@ class SpecialNaming:
     show_default=True,
     help="Optimize the images using pingo.",
 )
+@options.png_tagging
 @options.exiftool_path
 @options.pingo_path
 @options.use_bracket_type
@@ -132,6 +133,7 @@ def prepare_releases(
     image_quality: Optional[str],
     do_exif_tagging: bool,
     do_img_optimize: bool,
+    do_png_tagging: bool,
     exiftool_path: str,
     pingo_path: str,
     bracket_type: Literal["square", "round", "curly"],
@@ -286,7 +288,13 @@ def prepare_releases(
         optimize_images(pingo_exe, path_or_archive)
     if exiftool_exe is not None and do_exif_tagging:
         console.info("Tagging images with exif metadata...")
-        inject_metadata(exiftool_exe, path_or_archive, image_titling, rls_email)
+        inject_metadata(
+            exiftool_exe,
+            path_or_archive,
+            image_titling,
+            rls_email,
+            enable_png_tag=do_png_tagging,
+        )
 
 
 @click.command(
@@ -345,6 +353,7 @@ def prepare_releases(
     show_default=True,
     help="Optimize the images using pingo.",
 )
+@options.png_tagging
 @options.exiftool_path
 @options.pingo_path
 @options.use_bracket_type
@@ -366,6 +375,7 @@ def prepare_releases_chapter(
     image_quality: Optional[str],
     do_exif_tagging: bool,
     do_img_optimize: bool,
+    do_png_tagging: bool,
     exiftool_path: str,
     pingo_path: str,
     bracket_type: Literal["square", "round", "curly"],
@@ -464,4 +474,10 @@ def prepare_releases_chapter(
         optimize_images(pingo_exe, path_or_archive)
     if exiftool_exe is not None and do_exif_tagging:
         console.info("Tagging images with exif metadata...")
-        inject_metadata(exiftool_exe, path_or_archive, image_titling, rls_email)
+        inject_metadata(
+            exiftool_exe,
+            path_or_archive,
+            image_titling,
+            rls_email,
+            enable_png_tag=do_png_tagging,
+        )